Create Investment portfolio analysis
I made a python program that analyzes investment portfolio of a certain commodity of a country. It basically takes a user input on which country to invest and which commodity. Then it analyzes and gives an output of the annual return, annual volatility, sharp ratio and maximum drawbacks of the country